+++ src/routes/endpointGroup/context/createGroupContext.ts
@@ -69,7 +69,9 @@ module.exports = function (logger, app: express.Express, spinalAPIMiddleware: IS
       await SpinalGraphService.setGraph(graph);
       
       const context = await groupManagerService.createGroupContext(req.body.contextName, "BmsEndpoint")
-      userGraph.addContext(context);
+      if(userGraph._server_id != graph._server_id){
+        await userGraph.addContext(context);
+      }
 
       res.status(200).json({
         name: context.getName().get(),