+++ src/routes/tickets/workflows/createWorkflow.ts
@@ -57,7 +57,7 @@ module.exports = function (
    *                 description: name of the workflow
    *               steps:
    *                 type: array
-   *                 description: optionnal default steps that will be created in the workflow process
+   *                 description: optionnal default steps that will be created in the workflow process, steps start at order 1
    *                 items:
    *                   type: object
    *                   required:
@@ -72,7 +72,7 @@ module.exports = function (
    *                       description: color of the step
    *                     order:
    *                       type: integer
-   *                       description: order of the step
+   *                       description: order of the step, starts at 1
    *     responses:
    *       200:
    *         description: Create Successfully
@@ -100,7 +100,7 @@ module.exports = function (
       const steps: ITicketStep[] = [];
       if (req.body.steps && Array.isArray(req.body.steps)) {
         for (const step of req.body.steps) {
-          if (step.name && step.order) {
+          if (step.name && typeof step.order === 'number') {
             steps.push({
               name: step.name,
               color: step.color || undefined,
@@ -114,7 +114,9 @@ module.exports = function (
         req.body.nameWorkflow,
         steps
       );
-      await userGraph.addContext(contextTicketNode);
+      if(userGraph._server_id != graph._server_id){
+        await userGraph.addContext(contextTicketNode);
+      }
       await awaitSync(contextTicketNode);
 
       return res.status(200).json({