+++ src/routes/tickets/workflows/deleteWorkflow.ts
@@ -29,6 +29,7 @@ import { ISpinalAPIMiddleware } from '../../../interfaces';
 import {
   getAllTicketProcess,
   getStepNodesFromProcess,
+  getTicketsFromStep,
   TICKET_CONTEXT_TYPE,
 } from 'spinal-service-ticket';
 import { loadAndValidateNode } from '../../../utilities/loadAndValidateNode';
@@ -46,7 +47,7 @@ module.exports = function (
    *       - bearerAuth:
    *         - read
    *     description: Delete a workflow
-   *     summary: this will delete an workflow but also delete the related Process and Steps but NOT the tickets
+   *     summary: this will delete an workflow but also delete the related Process and Steps, deleting the tickets is an optional choice
    *     tags:
    *       - Workflow & ticket
    *     parameters:
@@ -57,6 +58,12 @@ module.exports = function (
    *        schema:
    *          type: integer
    *          format: int64
+   *      - in : query
+   *        name : shouldDeleteTickets
+   *        description : if true all tickets in the workflow will be deleted
+   *        required : false
+   *        schema:
+   *          type: boolean
    *     responses:
    *       200:
    *         description: Delete Successfully
@@ -67,6 +74,7 @@ module.exports = function (
   app.delete('/api/v1/workflow/:id/delete', async (req, res) => {
     try {
       const profileId = getProfileId(req);
+      const shouldDeleteTickets = req.query.shouldDeleteTickets === 'true';
       const workflowNode: SpinalContext = await loadAndValidateNode(
         spinalAPIMiddleware,
         parseInt(req.params.id, 10),
@@ -80,7 +88,14 @@ module.exports = function (
           processNode,
           workflowNode
         );
+
         for (const stepNode of stepNodes) {
+          const tickets = await getTicketsFromStep(stepNode);
+          if (shouldDeleteTickets) {
+            for (const ticket of tickets) {
+              proms.push(ticket.removeFromGraph());
+            }
+          }
           proms.push(stepNode.removeFromGraph());
         }
         proms.push(processNode.removeFromGraph());