+++ src/utilities/getParentNodesInfo.ts
@@ -7,21 +7,37 @@ async function getParentNodesInfo(
   spinalAPIMiddleware: ISpinalAPIMiddleware,
   profileId: string,
   dynamicId: number,
-  relations?: string[]
+  relations?: string[],
+  contextId? : number
 )  : Promise<BasicNode[]>{
     const node: SpinalNode<any> = await spinalAPIMiddleware.load(dynamicId,profileId);
-    const children = await node.getParents(relations);
-    const childrenInfo: BasicNode[] = [];
-    for (const child of children) {
-      const info : BasicNode = {
-        dynamicId: child._server_id,
-        staticId: child.getId().get(),
-        name: child.getName().get(),
-        type: child.getType().get(),
-      };
-      childrenInfo.push(info);
+    let contextNode = undefined;
+
+    if(contextId){
+      contextNode = await spinalAPIMiddleware.load(contextId,profileId);
     }
-    return childrenInfo;
+
+    let parents: SpinalNode<any>[];
+
+    if (contextNode) {
+      // children = await node.getChildrenInContext(contextNode, )
+      parents = await node.getParentsInContext(contextNode, relations);
+    }
+    else {
+      parents = await node.getParents(relations);
+    }
+
+    const parentsInfo: BasicNode[] = [];
+    for (const parent of parents) {
+      const info : BasicNode = {
+        dynamicId: parent._server_id,
+        staticId: parent.getId().get(),
+        name: parent.getName().get(),
+        type: parent.getType().get(),
+      };
+      parentsInfo.push(info);
+    }
+    return parentsInfo;
 
 
 }